BETA-0.2

+++ src/gui.py
@@ -9,6 +9,7 @@ import compare
 from PyQt5.QtWidgets import *
 from PyQt5.QtCore import pyqtSlot
 from PyQt5.QtGui import *
+from debug import Debug
 
 class rootWindow(QWidget):
     
@@ -25,7 +26,7 @@ class rootWindow(QWidget):
         self.w = None
         
     def build(self):
-        self.win.setWindowTitle("Collection Manager BETA-0.1")
+        self.win.setWindowTitle("Collection Manager BETA-0.2")
         self.win.setGeometry(100, 100, self.width, self.height)
         
         self.label = QLabel(self.win)
@@ -40,30 +41,45 @@ class rootWindow(QWidget):
 
         self.dup_button = QPushButton('Remove duplicates', self.win)
         self.dup_button.setToolTip('Remove all duplicates')
-        self.dup_button.setGeometry(70, 10, 150, 30)
+        self.dup_button.setGeometry(70, 10, 130, 30)
         self.dup_button.clicked.connect(self.remove_dup)
 
-        self.exp_button = QPushButton('Export scan', self.win)
-        self.exp_button.setToolTip('Export your scan to JSON file')
-        self.exp_button.setGeometry(230, 10, 100, 30)
+        self.exp_button = QPushButton('Export CSV', self.win)
+        self.exp_button.setToolTip('Export your scan to CSV file')
+        self.exp_button.setGeometry(210, 10, 100, 30)
         self.exp_button.clicked.connect(self.export_scan)
         
-        self.imp_button = QPushButton('Import JSON', self.win)
-        self.imp_button.setToolTip('Import scan with JSON file')
-        self.imp_button.setGeometry(340, 10, 100, 30)
-        self.imp_button.clicked.connect(self.import_json)
+        self.imp_button = QPushButton('Clean the base', self.win)
+        self.imp_button.setToolTip('Delete removed file in the base')
+        self.imp_button.setGeometry(320, 10, 100, 30)
+        self.imp_button.clicked.connect(self.clean_the_base)
         
         self.exp_button = QPushButton('Help', self.win)
         self.exp_button.setToolTip('Get help')
         self.exp_button.setGeometry(self.width - 50, 10, 40, 30)
         self.exp_button.clicked.connect(self.open_help)
         
+        self.exp_button = QPushButton('⚙', self.win)
+        self.exp_button.setToolTip('Open settings')
+        self.exp_button.setGeometry(self.width - 90, 10, 30, 30)
+        self.exp_button.clicked.connect(self.open_settings)
+        
         self.createTable(0)
         
+    def open_settings(self):
+        
+        if self.os == "windows":
+            
+            os.startfile("config.json")
+            
+        elif self.os == "linux":
+            
+            os.system("gnome-open config.json")
+    
     def open_help(self):
         if self.w is None:
             self.w = helpWindow()
-            self.w.build()
+            self.w.build(self)
             self.w.show()
 
         else:
@@ -72,39 +88,47 @@ class rootWindow(QWidget):
 
         
     @pyqtSlot()
-    def import_json(self):
+    def clean_the_base(self):
         
-        text, ok = QInputDialog.getText(self, 'Scan', 'Scan folder :')
-        if not ok:
+        dlg = QMessageBox(self)
+        dlg.setWindowTitle("Clean the base")
+        dlg.setText("Are you sure you want to clean the base ?")
+        dlg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+        dlg.setIcon(QMessageBox.Question)
+        button = dlg.exec()
+
+        if button != QMessageBox.Yes:
             return
         
-        self.file_json = json.loads(open(text, "r").read())
-        self.file_id = 0
-        self.sha256 = []
-        self.duplicates = []
-        
-        self.tableWidget.setRowCount(len(self.file_json) + 1)
-        
-        for file in self.file_json:
+        with open("db.json", "r") as r:
+            db = json.loads(r.read())
+            missing = list()
             
-            self.file_id += 1
-            self.setFile(file)
+            for k in db["collection"].keys():
+                
+                if "missing" in db["collection"][k]:
+                    
+                    missing.append(k)
+            
+            for k in missing:
+                del db["collection"][k]
+            
+            with open("db.json", "w") as w:
+                w.write(json.dumps(db, indent=4))
         
     @pyqtSlot()
     def scan(self):
         
-        self.text, ok = QInputDialog.getText(self, 'Scan', 'Scan folder :')
-        if not ok:
-            return
+        self.scan_folder = QFileDialog.getExistingDirectory(self, 'Select Folder')
         
         threading.Thread(target=self.scan_files).start()
         
     def scan_files(self):
         
-        if self.text == "#folder":
-            self.text = self.config["folder"]
+        #if self.text == "#folder":
+        #    self.text = self.config["folder"]
         
-        self.file_json = scanner.scan_files(self.text, self.os, self.config)
+        self.file_json = scanner.scan_files(self.scan_folder, self.os, self.config)
         self.file_id = 0
         self.sha256 = []
         self.duplicates = []
@@ -117,7 +141,6 @@ class rootWindow(QWidget):
             self.file_id += 1
             self.setFile(self.file_json[file], self.file_id)
         
-        print("----- SHA")
         scanner.get_files_sha(self, self.file_json)
         
         if self.config["already_scanned"] != "":
@@ -129,11 +152,13 @@ class rootWindow(QWidget):
         for miss in missing:
             
             self.file_id += 1
+            missing[miss]["missing"] = True
             self.setFile(missing[miss], self.file_id, True)
             
         
         with open("db.json", "r") as r:
             db = json.loads(r.read())
+            self.file_json.update(missing)
             db["collection"] = self.file_json
             with open("db.json", "w") as w:
                 w.write(json.dumps(db, indent=4))
@@ -160,15 +185,23 @@ class rootWindow(QWidget):
     @pyqtSlot()
     def remove_dup(self):
         
+        dlg = QMessageBox(self)
+        dlg.setWindowTitle("Remove duplicates")
+        dlg.setText("Are you sure you want to permanently delete files ?")
+        dlg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+        dlg.setIcon(QMessageBox.Question)
+        button = dlg.exec()
+
+        if button != QMessageBox.Yes:
+            return
+        
         if len(self.duplicates) > 0:
 
             for file_id in self.duplicates:
 
                 cell = self.tableWidget.item(file_id, 5)
-                print("remove > " + cell.text())
                 os.remove(cell.text())
-                print(self.duplicates)
-                print(file_id)
+                Debug.Log("Removed > " + cell.text())
 
         else:
 
@@ -179,11 +212,9 @@ class rootWindow(QWidget):
 
     def setFile(self, meta, file_id, missing = False):
         
-        print(meta)
-        print("----= " + str(file_id))
+        Debug.Log("Set file : " + str(meta))
                 
         self.tableWidget.setItem(file_id, 0, QTableWidgetItem(meta["name"]))
-        print(meta["name"])
         self.tableWidget.setItem(file_id, 1, QTableWidgetItem(meta["date"]))
         self.tableWidget.setItem(file_id, 2, QTableWidgetItem(meta["d_size"]))
         self.tableWidget.setItem(file_id, 3, QTableWidgetItem(meta["disk"]))
@@ -209,7 +240,6 @@ class rootWindow(QWidget):
                     
             else:
                 for i in range(7):
-                    print("---- " + str(i))
                     self.tableWidget.item(file_id, i).setBackground(QColor(255,100,100))
                 
         else:
@@ -217,7 +247,6 @@ class rootWindow(QWidget):
             self.tableWidget.setItem(file_id, 6, QTableWidgetItem("..."))
             
             for c in range(7):
-                print(c)
                 self.tableWidget.item(file_id, c).setBackground(QColor(255,200,150))
                 
         return True
@@ -255,9 +284,9 @@ class helpWindow(QWidget):
         self.height = 200
         self.width = 400
         
-    def build(self):
-        self.setWindowTitle("Collection Manager BETA-0.1")
-        self.setGeometry(100, 100, self.width, self.height)
+    def build(self, win):
+        self.setWindowTitle("Collection Manager BETA-0.2")
+        self.setGeometry(win.win.pos().x() + (win.width/2 - self.width/2), win.win.pos().y() + (win.height/2 - self.height/2), self.width, self.height)
         
         self.label = QLabel(self)
         self.label.setText("Collection Manager")
@@ -265,7 +294,7 @@ class helpWindow(QWidget):
         self.label.setStyleSheet("font-size: 22px")
         
         self.label = QLabel(self)
-        self.label.setText("Version : BETA-0.1")
+        self.label.setText("Version : BETA-0.2")
         self.label.setGeometry(round(self.width/2 - 190/2), 40, 190, 30)
         
         self.label = QLabel(self)
@@ -277,9 +306,13 @@ class helpWindow(QWidget):
         self.label.setGeometry(round(self.width/2 - 190/2), 70, 190, 30)
         
 def start(os_system):
+    global Debug
+    
+    Debug = Debug(json.loads(open("config.json").read())["log"])
+    
     app = QApplication(sys.argv)
     root = QWidget()
-    print(os_system)
+    Debug.Info(os_system)
     rootWin = rootWindow(root, os_system)
     rootWin.build()
